requirements.txt
Flask==1.1.2
notion==0.0.25
python-dotenv==0.14.0
tweepy==3.9.0
